// src/common/types.ts
export type Maybe<T> = T | null
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K]
}
export type MakeOptional<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]?: Maybe<T[SubKey]> }
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]: Maybe<T[SubKey]> }
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  Upload: any
}
// ...
export enum OperatorKindEnum {
  Gte = "GTE",
  Lte = "LTE",
  Ne = "NE",
  Like = "LIKE"
}

export type OperatorOptions = {
  kind?: Maybe<OperatorKindEnum>
  field?: Maybe<Scalars["String"]>
  value?: Maybe<Scalars["String"]>
}

export type PageLimitPair = {
  __typename?: "PageLimitPair"
  page?: Maybe<Scalars["Int"]>
  limit?: Maybe<Scalars["Int"]>
}

export type PageMetadata = {
  __typename?: "PageMetadata"
  totalCount?: Maybe<Scalars["Int"]>
}

export type PageQueryOptions = {
  paginate?: Maybe<PaginateOptions>
  slice?: Maybe<SliceOptions>
  sort?: Maybe<Array<Maybe<SortOptions>>>
  operators?: Maybe<Array<Maybe<OperatorOptions>>>
  search?: Maybe<SearchOptions>
}

export type PaginateOptions = {
  page?: Maybe<Scalars["Int"]>
  limit?: Maybe<Scalars["Int"]>
}

export type PaginationLinks = {
  __typename?: "PaginationLinks"
  first?: Maybe<PageLimitPair>
  prev?: Maybe<PageLimitPair>
  next?: Maybe<PageLimitPair>
  last?: Maybe<PageLimitPair>
}
// ...
export type UsersQuery = {
  users: UsersPage;
};

export type UserQuery = {
  user: User;
}

export type TodosQuery = {
  todos: TodosPage;
}
// ...
export type SearchOptions = {
  q?: Maybe<Scalars["String"]>
}

export type SliceOptions = {
  start?: Maybe<Scalars["Int"]>
  end?: Maybe<Scalars["Int"]>
  limit?: Maybe<Scalars["Int"]>
}

export type SortOptions = {
  field?: Maybe<Scalars["String"]>
  order?: Maybe<SortOrderEnum>
}

export enum SortOrderEnum {
  Asc = "ASC",
  Desc = "DESC"
}

export type Todo = {
  __typename?: "Todo"
  id: Scalars["ID"]
  title: Scalars["String"]
  completed: Scalars["Boolean"]
  user: User
}

export type TodosPage = {
  __typename?: "TodosPage"
  data?: Maybe<Array<Maybe<Todo>>>
  links?: Maybe<PaginationLinks>
  meta?: Maybe<PageMetadata>
}
// ...
export type User = {
  __typename?: "User"
  id?: Maybe<Scalars["ID"]>
  name?: Maybe<Scalars["String"]>
  username?: Maybe<Scalars["String"]>
  email?: Maybe<Scalars["String"]>
  address?: Maybe<Address>
  phone?: Maybe<Scalars["String"]>
  website?: Maybe<Scalars["String"]>
  company?: Maybe<Company>
  posts?: Maybe<PostsPage>
  albums?: Maybe<AlbumsPage>
  todos?: Maybe<TodosPage>
}

export type UserPostsArgs = {
  options?: Maybe<PageQueryOptions>
}

export type UserAlbumsArgs = {
  options?: Maybe<PageQueryOptions>
}

export type UserTodosArgs = {
  options?: Maybe<PageQueryOptions>
}

export type UsersPage = {
  __typename?: "UsersPage"
  data?: Maybe<Array<Maybe<User>>>
  links?: Maybe<PaginationLinks>
  meta?: Maybe<PageMetadata>
}